rust/workflow: add trinary_input_string
Generic workflow driving the trinary input string component. Will also
be used to enter the multisig account name on the device.
