+++ src/rust/bitbox02-rust/src/workflow.rs
@@ -19,5 +19,6 @@ pub mod pairing;
 pub mod password;
 pub mod sdcard;
 pub mod status;
+pub mod trinary_input_string;
 pub mod unlock;
 pub mod verify_message;
